+++ cryptoparser/common/parse.py
@@ -9,10 +9,6 @@ import time
 
 import attr
 
-import dateutil
-import dateutil.parser
-import dateutil.tz
-
 from cryptodatahub.common.exception import InvalidValue
 from cryptodatahub.common.types import CryptoDataEnumBase, CryptoDataEnumCodedBase
 
@@ -466,9 +462,15 @@ class ParserText(ParserBase):
     def parse_date_time(self, name):
         try:
             value = self._parsable[self._parsed_length:]
-            date_time = dateutil.parser.parse(value.decode(self._encoding))
-        except ValueError as e:
-            raise InvalidValue(value, type(self), 'value') from e
+            value_str = value.decode(self._encoding)
+
+            date_time = datetime.datetime.strptime(value_str, '%a, %d %b %Y %H:%M:%S %z')
+        except ValueError:
+            try:
+                date_time = datetime.datetime.strptime(value_str, '%a, %d %b %Y %H:%M:%S GMT')
+                date_time = date_time.replace(tzinfo=datetime.timezone.utc)
+            except ValueError as e:
+                raise InvalidValue(value, type(self), 'value') from e
 
         self._parsed_values[name] = date_time
         self._parsed_length = len(self._parsable)
@@ -500,7 +502,7 @@ class ParserBinary(ParserBase):
             if milliseconds:
                 millis = value % 1000
                 value //= 1000
-            value = datetime.datetime.fromtimestamp(0x00000000ffffffff & value, dateutil.tz.UTC)
+            value = datetime.datetime.fromtimestamp(0x00000000ffffffff & value, datetime.timezone.utc)
             if milliseconds:
                 value += datetime.timedelta(milliseconds=millis)
             self._parsed_values[name] = value